requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.1.0
Flask-WTF==1.2.2
Flask-Login==0.6.3
python-dotenv==1.0.1
WTForms==3.2.1
